// chrome/src/models/das-events.enum.ts
export enum DasEventsEnum {
    CURRENT_TAB_ID = 'CURRENT_TAB_ID',
    PREVIOUS_TAB_ID = 'PREVIOUS_TAB_ID',
    TEMPORARY_WHITELIST = 'TEMPORARY_WHITELIST',
    SUSPENDED_WINDOW_TITLE_ICON = '💤'
}

export enum DasEventsNamesEnum {
    DISCARD_ONE = 'DISCARD_ONE',
    DISCARD_ALL = 'DISCARD_ALL',
    RELOAD_ALL = 'RELOAD_ALL',
    DISCARD_SELECTED_TABS = 'DISCARD_SELECTED_TABS',
    RELOAD_SELECTED_TABS = 'RELOAD_SELECTED_TABS',
    WHITELIST_TAB = 'WHITELIST_TAB',
    REMOVE_WHITELIST_TAB = 'REMOVE_WHITELIST_TAB',
    TEMP_WHITELIST_TAB = 'TEMP_WHITELIST_TAB',
    TEMP_UNDO_WHITELIST_TAB = 'TEMP_UNDO_WHITELIST_TAB'
}

export enum DasEventsInternals {
    REQUEST_CURRENT_TAB_INFO = 'REQUEST_CURRENT_TAB_INFO',
    RESET_TAB_TIMERS = 'RESET_TIMERS',
    UPDATE_CONTEXT_MENU_ITEMS = 'UPDATE_CONTEXT_MENU_ITEMS'
}

export enum DasEventsTabStatuses {
    NORMAL = 'NORMAL',
    SPECIAL = 'SPECIAL',
    DISCARDED = 'DISCARDED',
    WHITELISTED = 'WHITELISTED',
    TEMP_WHITELISTED = 'TEMP_WHITELISTED',
    PINNED = 'PINNED',
    AUDIBLE = 'AUDIBLE',
    NEVER = 'NEVER',
    CHARGING = 'CHARGING',
    NO_CONNECTIVITY = 'NO_CONNECTIVITY',
    UNKNOWN = 'UNKNOWN'
}